Added universal.bashrc and moved "bin" -> "python"
Also added some very small utilities: bcols and brealpath.

"bin" directory was renamed to "python" in preparation for having
multiple types of scripts that will be installed differently by the
future installer. There may end up being a "BASH" directory, but
universal.bashrc will probably not be in it because it's not something
that would be called the same way as other utilities and it ought to be
specially handled during installation.