Remove cargo install support from Windows
This didn't seem to be working right on Windows. It was using the
installation method that cargo suggests for macOS and linux there too
and it seems like that actually does work _if_ the Visual Studio build
tools are installed but causes some crazy linker errors otherwise. It
would be nice to fix this at some point, but for now I'm just going to
turn it off.